Fix broken change detection for mutable fields (e.g. Bitmask) after fault-in through polymorphic reference whose declared class does not have that field.
Set the #_class: of the proxy *before* assigning the proxiedObject, so that #initializeCopyObject will operate on the full list of fields for the class of the object, even if the reference was declared using a superclass.

Also raise an error if the retrieved class does not #includesBehavior: the declared/expected class (e.g. accidentally requesting deferredObjectOfClass: SubPerson withID: <aMidPersonID>).